Create go structures to hold OpenControls data
This creates both a parser and the structures that will ultimately hold
the opencontrol data. These are currently optimized for search and
iteration, but we'll ultimately need to transform this into appropraite
json/yaml representations.

Currently the code just parses it and outputs what it parsed. The text
implementation is currently stubbed and needs to be fixed.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>