Add unit tests and handle narratives for all control types
This adds simple unit tests for control and family parsing. Thus
documenting a little more what's expected from parsing those controls.

This also handles narratives a little better. Note that currently there
are placeholders for the narratives so we are able to test them. We
should replace these with actual texts and pass these through the
appropraite functions. This will require us to standardize on where in
the spreadsheet should these texts go.

Finally, this adds the unit tests run to the workflow

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>